pyproject.toml
[tool.poetry]
name = "nome"
version = "0.1.0"
description = ""
authors = ["alessiogandelli <alessiogandelli99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
pandas = "^2.1.0"
matplotlib = "^3.7.2"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
seaborn = "^0.13.0"
jsonlines = "^4.0.0"
bertopic = "^0.16.0"
openai = "^1.5.0"
gsdmm = {git = "https://github.com/rwalk/gsdmm.git"}
contractions = "^0.1.73"
tweets-to-topic-network = "^0.1.3"
flask = "^3.0.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
